api/requirements-dev.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.10.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-mock==3.14.0
pytest-cov==4.1.0
requests-mock==1.12.0
playwright==1.45.0
pytest-playwright==0.5.1